requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19
python-dotenv==1.0.1

# AI / Vision
anthropic==0.42.0

# OCR
pytesseract==0.3.13
easyocr==1.7.2

# Document handling
PyMuPDF==1.25.1
Pillow==11.1.0
python-docx==1.1.2

# Database
SQLAlchemy==2.0.36

# Utilities
pydantic==2.10.4
pydantic-settings==2.7.1
aiofiles==24.1.0
